+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added support for Persistent History Tracking.
 - Added support for batch inserts.
+- Added support for async fetch requests.
 - Added `materialize()` method on `NSManagedObject`.
 - `Collection.fetchFaultedObjects()` deprecated, use `Collection.materializeFaultedObjects()` instead.
 - Added support for the new Xcode 11 SPM.